minor correction.
git-svn-id: svn://vsl.cis.udel.edu/civl/trunk@2139 fb995dde-84ed-4084-dfe6-e5aef3e2452c

+++ examples/mpi/collective/allreduce.c
@@ -24,7 +24,7 @@ int main(int argc, char **argv){
     if(mype%2)
         ierr = MPI_Allreduce(&val, &sum, knt, MPI_DOUBLE, MPI_SUM, WCOMM);
     else
-        ierr = MPI_Allreduce(&val, &sum, knt, MPI_INT, MPI_MAX, WCOMM);
+        ierr = MPI_Allreduce(&val, &sum, knt, MPI_DOUBLE, MPI_MAX, WCOMM);
 #else
     ierr = MPI_Allreduce(&val, &sum, knt, MPI_DOUBLE, MPI_SUM, WCOMM);
 #endif
